mac80211: introduce modparam to configure association timeout
Current MLME logic in mac80211 only waits one beacon interval before
failing the association. In conditions of high beacon loss triggers this
failure and requires a restart from the authentication stage. This
results in poor association times.

This change enables configuration of the association timeout by adjusting
the beacon intervals to wait for a beacon.